+++ UVtools.Core/FileFormats/AnycubicZipFile.cs
@@ -6,6 +6,9 @@
  *  of this license document, but changing it is not allowed.
  */
 
+using BinarySerialization;
+using Emgu.CV;
+using Emgu.CV.CvEnum;
 using System;
 using System.Drawing;
 using System.IO;
@@ -13,14 +16,17 @@ using System.IO.Compression;
 using System.Runtime.InteropServices;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Threading.Tasks;
 using UVtools.Core.Converters;
+using UVtools.Core.EmguCV;
 using UVtools.Core.Extensions;
 using UVtools.Core.Layers;
 using UVtools.Core.Operations;
+using static UVtools.Core.FileFormats.AnycubicFile;
 
 namespace UVtools.Core.FileFormats;
 
-public sealed class PWSZFile : FileFormat
+public sealed class AnycubicZipFile : FileFormat
 {
     #region Sub classes
 
@@ -280,6 +286,117 @@ public sealed class PWSZFile : FileFormat
             OS = RuntimeInformation.RuntimeIdentifier;
         }
     }
+
+    public sealed class SceneLayerDef
+    {
+        [FieldOrder(0)] public float Height { get; set; }
+
+        [FieldOrder(1)] public float Area { get; set; }
+
+        [FieldOrder(2)] public float XMin { get; set; }
+
+        [FieldOrder(3)] public float YMin { get; set; }
+
+        [FieldOrder(4)] public float XMax { get; set; }
+
+        [FieldOrder(5)] public float YMax { get; set; }
+
+        [FieldOrder(6)] public uint ContourCount { get; set; }
+
+        [FieldOrder(7)] public float MaxContourArea { get; set; }
+
+        [FieldOrder(8)] [FieldCount(8)] public uint[] Padding { get; set; } = new uint[8];
+
+        public override string ToString()
+        {
+            return
+                $"{nameof(Height)}: {Height}, {nameof(Area)}: {Area}, {nameof(XMin)}: {XMin}, {nameof(YMin)}: {YMin}, {nameof(XMax)}: {XMax}, {nameof(YMax)}: {YMax}, {nameof(ContourCount)}: {ContourCount}, {nameof(MaxContourArea)}: {MaxContourArea}, {nameof(Padding)}: {Padding}";
+        }
+    }
+
+    public sealed class SceneManifest
+    {
+        public const string MAGIC = "ANYCUBIC-PWSZ";
+
+        [FieldOrder(0)]
+        [FieldLength(16)]
+        [SerializeAs(SerializedType.TerminatedString)]
+        public string Magic { get; set; } = MAGIC;
+
+        [FieldOrder(1)]
+        [FieldLength(64)]
+        [SerializeAs(SerializedType.TerminatedString)]
+        public string Software { get; set; } = About.SoftwareWithVersionArch;
+
+        /// <summary>
+        /// 1: Pure Binary<br/>
+        /// 2: FPGA Debug<br/>
+        /// 3: FPGA Release. Firmware set to 3 (Default)
+        /// </summary>
+        [FieldOrder(2)] public uint BinaryType { get; set; } = 3;
+
+        [FieldOrder(3)] public uint Version { get; set; } = 1;
+
+        [FieldOrder(4)] public uint SliceType { get; set; }
+
+        /// <summary>
+        /// 0: mm<br/>
+        /// 1: cm<br/>
+        /// 2: m
+        /// </summary>
+        [FieldOrder(5)] public uint ModelUnit { get; set; }
+
+        [FieldOrder(6)] public float PointRatio { get; set; } = 1f;
+
+        [FieldOrder(7)] public uint LayerCount { get; set; }
+
+        [FieldOrder(8)] public float XMin { get; set; }
+
+        [FieldOrder(9)] public float YMin { get; set; }
+
+        [FieldOrder(10)] public float ZMin { get; set; }
+
+        [FieldOrder(11)] public float XMax { get; set; } = 1;
+
+        [FieldOrder(12)] public float YMax { get; set; }
+
+        [FieldOrder(13)] public float ZMax { get; set; }
+
+        /// <summary>
+        /// Some status flags of the scene model
+        /// </summary>
+        [FieldOrder(14)] public uint ModelStats { get; set; }
+
+        [FieldOrder(15)] [FieldCount(64)] public uint[] Padding { get; set; } = new uint[64];
+
+        [FieldOrder(16)] [FieldLength(4)] public string Separator { get; set; } = "<---";
+
+        [FieldOrder(17)] public uint LayerDefCount { get; set; }
+
+        [FieldOrder(18)][FieldCount(nameof(LayerDefCount))] public SceneLayerDef[] LayersDef { get; set; } = Array.Empty<SceneLayerDef>();
+
+        [FieldOrder(19)] [FieldLength(4)] public string EndMarker { get; set; } = "--->";
+
+        public void Update(FileFormat slicerFile)
+        {
+            Software = About.SoftwareWithVersionArch;
+            var rect = slicerFile.BoundingRectangleMillimeters;
+            rect.Offset(slicerFile.DisplayWidth / -2f, slicerFile.DisplayHeight / -2f);
+            XMin = (float)Math.Round(rect.X, 4);
+            YMin = (float)Math.Round(rect.Y, 4);
+            XMax = (float)Math.Round(rect.Right, 4);
+            YMax = (float)Math.Round(rect.Bottom, 4);
+
+            ZMin = 0;
+            ZMax = slicerFile.PrintHeight;
+        }
+
+        public override string ToString()
+        {
+            return
+                $"{nameof(Magic)}: {Magic}, {nameof(Software)}: {Software}, {nameof(BinaryType)}: {BinaryType}, {nameof(Version)}: {Version}, {nameof(SliceType)}: {SliceType}, {nameof(ModelUnit)}: {ModelUnit}, {nameof(PointRatio)}: {PointRatio}, {nameof(LayerCount)}: {LayerCount}, {nameof(XMin)}: {XMin}, {nameof(YMin)}: {YMin}, {nameof(ZMin)}: {ZMin}, {nameof(XMax)}: {XMax}, {nameof(YMax)}: {YMax}, {nameof(ZMax)}: {ZMax}, {nameof(ModelStats)}: {ModelStats}, {nameof(Padding)}: {Padding}, {nameof(Separator)}: {Separator}, {nameof(LayerDefCount)}: {LayerDefCount}, {nameof(LayersDef)}: {LayersDef}, {nameof(EndMarker)}: {EndMarker}";
+        }
+    }
     #endregion
 
     #region Constants
@@ -295,6 +412,7 @@ public sealed class PWSZFile : FileFormat
     public PrintInfoManifest PrintInfoSettings { get; set; } = new ();
     public LayersControllerManifest LayersSettings { get; set; } = new ();
     public SoftwareInfoManifest SoftwareInfoSettings { get; set; } = new ();
+    public SceneManifest SceneSettings { get; set; } = new ();
 
     public override FileFormatType FileType => FileFormatType.Archive;
 
@@ -330,9 +448,23 @@ public sealed class PWSZFile : FileFormat
     public override string ConvertMenuGroup => "Anycubic Photon Workshop";
 
     public override FileExtension[] FileExtensions { get; } = {
-        new(typeof(PWSZFile), "pm7", "Photon Mono M7 (PM7)"),
-        new(typeof(PWSZFile), "pwsz", "Photon Mono M7 Pro (PWSZ)")
+        new(typeof(AnycubicZipFile), "pm7", "Photon Mono M7 (PM7)"),
+        new(typeof(AnycubicZipFile), "pwsz", "Photon Mono M7 Pro (PWSZ)")
     };
+
+    /*public override uint[] AvailableVersions { get; } = { 1 };
+
+    public override uint DefaultVersion => 1;
+
+    public override uint Version
+    {
+        get => SceneSettings.Version;
+        set
+        {
+            base.Version = value;
+            SceneSettings.Version = base.Version;
+        }
+    }*/
 
     public override uint ResolutionX
     {
@@ -364,6 +496,12 @@ public sealed class PWSZFile : FileFormat
         set => base.MachineZ = Settings.MachineType.MachineZ = value;
     }
 
+    public override uint LayerCount
+    {
+        get => base.LayerCount;
+        set => base.LayerCount = SceneSettings.LayerCount = SceneSettings.LayerDefCount = base.LayerCount;
+    }
+
     public override string MachineName
     {
         get => Settings.MachineType.Name;
@@ -382,17 +520,93 @@ public sealed class PWSZFile : FileFormat
 
 
     public override object[] Configs => new object[] { 
-        Settings, PrintInfoSettings, SoftwareInfoSettings
+        Settings, PrintInfoSettings, SoftwareInfoSettings, SceneSettings
     };
 
     #endregion
 
     #region Constructor
-    public PWSZFile()
+    public AnycubicZipFile()
     { }
     #endregion
 
     #region Methods
+
+    // PW0
+    private Mat DecodeLayerRle(byte[] encodedRle)
+    {
+        var mat = CreateMat();
+        var imageLength = mat.GetLength();
+
+        int pixelPos = 0;
+        for (int i = 0; i < encodedRle.Length; i++)
+        {
+            byte b = encodedRle[i];
+            int code = b >> 4;
+            int repeat = b & 0xf;
+            byte color;
+            switch (code)
+            {
+                case 0x0:
+                    color = 0;
+                    i++;
+                    //reps = reps * 256 + EncodedRle[i];
+                    if (i >= encodedRle.Length)
+                    {
+                        repeat = imageLength - pixelPos;
+                        break;
+                    }
+
+                    repeat = (repeat << 8) + encodedRle[i];
+                    break;
+                case 0xf:
+                    color = 255;
+                    i++;
+                    //reps = reps * 256 + EncodedRle[i];
+                    if (i >= encodedRle.Length)
+                    {
+                        repeat = imageLength - pixelPos;
+                        break;
+                    }
+
+                    repeat = (repeat << 8) + encodedRle[i];
+                    break;
+                default:
+                    color = (byte)((code << 4) | code);
+                    if (i >= encodedRle.Length)
+                    {
+                        repeat = imageLength - pixelPos;
+                    }
+                    break;
+            }
+
+            //color &= 0xff;
+
+            if (pixelPos + repeat > imageLength)
+            {
+                mat.Dispose();
+                throw new FileLoadException($"Image ran off the end: {pixelPos} + {repeat} = {pixelPos + repeat}, expecting: {imageLength}");
+            }
+
+            // We only need to set the non-zero pixels
+            mat.FillSpan(ref pixelPos, repeat, color);
+
+
+            if (pixelPos == imageLength)
+            {
+                //i++;
+                break;
+            }
+        }
+
+        if (pixelPos > 0 && pixelPos != imageLength)
+        {
+            mat.Dispose();
+            throw new FileLoadException($"Image ended short: {pixelPos}, expecting: {imageLength}");
+        }
+
+        return mat;
+    }
 
     protected override void OnBeforeEncode(bool isPartialEncode)
     {
@@ -439,6 +653,8 @@ public sealed class PWSZFile : FileFormat
 
             minHeight += relativeZ;
         }
+
+        SceneSettings.Update(this);
     }
 
     protected override void EncodeInternally(OperationProgress progress)
@@ -446,12 +662,54 @@ public sealed class PWSZFile : FileFormat
         using var outputFile = ZipFile.Open(TemporaryOutputFileFullPath, ZipArchiveMode.Create);
 
         EncodeAllThumbnailsInZip(outputFile, "preview_images/preview_{0}.png", progress);
-        //EncodeLayersInZip(outputFile, IndexStartNumber.One, progress);
-
+        
         outputFile.CreateEntryFromSerializeJson(SettingsFileName, Settings, ZipArchiveMode.Create, JsonExtensions.SettingsIndent);
         outputFile.CreateEntryFromSerializeJson(PrintInfoFileName, PrintInfoSettings, ZipArchiveMode.Create, JsonExtensions.SettingsIndent);
         outputFile.CreateEntryFromSerializeJson(LayersFileName, LayersSettings, ZipArchiveMode.Create, JsonExtensions.SettingsIndent);
         outputFile.CreateEntryFromSerializeJson(SoftwareInfoFileName, SoftwareInfoSettings, ZipArchiveMode.Create, JsonExtensions.SettingsIndent);
+
+        SceneSettings.LayersDef = new SceneLayerDef[LayerCount];
+        var pixelArea = PixelArea;
+
+        progress.Reset(OperationProgress.StatusEncodeLayers, LayerCount);
+        foreach (var batch in BatchLayersIndexes())
+        {
+            Parallel.ForEach(batch, CoreSettings.GetParallelOptions(progress), layerIndex =>
+            {
+                progress.PauseIfRequested();
+
+                var layer = this[layerIndex];
+                using var mat = layer.LayerMatModelBoundingRectangle;
+
+                var rect = layer.BoundingRectangleMillimeters;
+                rect.Offset(DisplayWidth / -2f, DisplayHeight / -2f);
+
+                using var contours = new EmguContours(mat.RoiMat, RetrType.External);
+
+                SceneSettings.LayersDef[layerIndex] = new SceneLayerDef
+                {
+                    Height = this[layerIndex].PositionZ,
+                    Area = layer.GetArea(),
+                    XMin = (float)Math.Round(rect.X, 4),
+                    YMin = (float)Math.Round(rect.Y, 4),
+                    XMax = (float)Math.Round(rect.Right, 4),
+                    YMax = (float)Math.Round(rect.Bottom, 4),
+                    ContourCount = (uint)contours.ExternalContoursCount,
+                    MaxContourArea = (float)Math.Round(contours.MaxSolidArea * pixelArea, 4)
+                };
+
+                progress.LockAndIncrement();
+            });
+
+            foreach (var layerIndex in batch)
+            {
+                progress.PauseOrCancelIfRequested();
+            }
+        }
+
+        using var memoryStream = new MemoryStream();
+        Helpers.Serializer.Serialize(memoryStream, SceneSettings);
+        outputFile.CreateEntryFromContent(SceneFileName, memoryStream, ZipArchiveMode.Create);
     }
 
     protected override void DecodeInternally(OperationProgress progress)
@@ -527,9 +785,25 @@ public sealed class PWSZFile : FileFormat
             }
         }
 
+        entry = inputFile.GetEntry(SceneFileName);
+        if (entry is not null)
+        {
+            try
+            {
+                using var stream = entry.Open();
+                SceneSettings = Helpers.Deserialize<SceneManifest>(stream);
+            }
+            catch (Exception e)
+            {
+                Clear();
+                throw new FileLoadException($"Unable to deserialize '{entry.Name}'\n{e}", FileFullPath);
+            }
+        }
+
         DecodeAllThumbnailsFromZip(inputFile, progress, "preview_");
 
         Init((uint)LayersSettings.Count);
+        progress.Reset(OperationProgress.StatusDecodeLayers, LayerCount);
 
         float positionZ = 0;
         for (uint layerIndex = 0; layerIndex < LayerCount; layerIndex++)
@@ -544,7 +818,30 @@ public sealed class PWSZFile : FileFormat
                 LiftSpeed = SpeedConverter.Convert(LayersSettings.Layers[layerIndex].LiftSpeed, FormatSpeedUnit, CoreSpeedUnit)
             };
         }
-        //DecodeLayersFromZip(inputFile, IndexStartNumber.One, progress);
+
+        if (DecodeType == FileDecodeType.Full)
+        {
+            Parallel.For(0, LayerCount, CoreSettings.GetParallelOptions(progress), layerIndex =>
+            {
+                progress.PauseIfRequested();
+                byte[] encodedRle;
+                lock (Mutex)
+                {
+                    entry = inputFile.GetEntry($"layer_images/layer_{layerIndex}.pwszImg");
+                    if (entry is null)
+                    {
+                        Clear();
+                        throw new FileLoadException($"Layer image {layerIndex} is missing in the file.", FileFullPath);
+                    }
+
+                    using var stream = entry.Open();
+                    encodedRle = stream.ToArray();
+                }
+
+                this[layerIndex].LayerMat = DecodeLayerRle(encodedRle);
+                progress.LockAndIncrement();
+            });
+        }
 
         UpdateGlobalPropertiesFromLayers();
     }