+++ UVtools.Core/FileFormats/CrealityCXDLPv1File.cs
@@ -22,7 +22,7 @@ using UVtools.Core.Operations;
 
 namespace UVtools.Core.FileFormats;
 
-public sealed class CXDLPv1File : FileFormat
+public sealed class CrealityCXDLPv1File : FileFormat
 {
     #region Constants
     private const byte HEADER_SIZE = 9; // CXSW3DV2
@@ -311,10 +311,10 @@ public sealed class CXDLPv1File : FileFormat
 
     public override FileFormatType FileType => FileFormatType.Binary;
 
-    public override string ConvertMenuGroup => "CXDLP";
+    public override string ConvertMenuGroup => "Creality CXDLP";
 
     public override FileExtension[] FileExtensions { get; } = {
-        new(typeof(CXDLPv1File), "v1.cxdlp", "Creality CXDLPv1"),
+        new(typeof(CrealityCXDLPv1File), "v1.cxdlp", "Creality CXDLPv1"),
     };
 
     public override PrintParameterModifier[] PrintParameterModifiers { get; } =